Moving load functions to common.h
